partially implement hand manager
